Remove snapshot_darwin and snapshop_linux
