+++ snapshot.go
@@ -3,7 +3,9 @@ package restic
 import (
 	"fmt"
 	"os"
+	"os/user"
 	"path/filepath"
+	"strconv"
 	"time"
 
 	"github.com/restic/restic/backend"
@@ -64,4 +66,26 @@ func (sn Snapshot) String() string {
 
 func (sn Snapshot) ID() backend.ID {
 	return sn.id
+}
+
+func (sn *Snapshot) fillUserInfo() error {
+	usr, err := user.Current()
+	if err != nil {
+		return err
+	}
+
+	sn.Username = usr.Username
+	uid, err := strconv.ParseInt(usr.Uid, 10, 32)
+	if err != nil {
+		return err
+	}
+	sn.UID = uint32(uid)
+
+	gid, err := strconv.ParseInt(usr.Gid, 10, 32)
+	if err != nil {
+		return err
+	}
+	sn.GID = uint32(gid)
+
+	return nil
 }
